+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * export `param_converter` from core `flask_unchained` package
 * fix discovery of user-app `tests._unchained_config`
 * improve the output of commands that display their information in a table
+* improve the output of custom sqlalchemy types in generated migrations files
 
 ## 0.5.1 (2018/07/25)